requirements.txt
asyncio
azure-identity
azure-core
azure-ai-projects
python-dotenv
requests
markdown-pdf
fpdf2
aiohttp